Update Hungarian language file
Sorry for being unofficial, but I prepared the naturalization in Hungarian. Although I didn't manage to insert it into the program's installer (I don't know how it works). Anyway, I put the .mo file in the installation location and modified the xml, but it didn't change to any language other than English.